requirements.txt
filemagic==1.6
